DocumentationProvider: Move into documentation package
